Revamp the CSS styling based on  the box model
Re-styled the pages by adding a few classes to the
HTML and using them in the CSS.
* Styles.css needs tidying up
* Styles.css needs reworking so the Carbonara example
includes the steps and comments within a single box.
* Need to figure out why background-color in the
container class bleeds out into the whole page.
* The side-by-side styling is not entirely comprehensible
it just worked with the inline-flex display type and
i left it as such. Not happy about that.
* VSCode claims tere are EIGHT mistakes. sort them out.